jvm memory opt

+++ pixels-cache/src/test/java/cn/edu/ruc/iir/pixels/cache/TestPixelsCacheReader.java
@@ -24,7 +24,7 @@ public class TestPixelsCacheReader
                     .setMQAppend(true)
                     .build();
 
-            byte[] content = cacheReader.get(1, 1, 1);
+            byte[] content = cacheReader.get(1, (short) 1, (short) 1);
         }
         catch (Exception e) {
             e.printStackTrace();
